Reverting change to the partial updates.
