+++ flows/boundary.py
@@ -22,13 +22,12 @@ from cpr_sdk.ssm import get_aws_ssm_param
 from prefect import get_run_logger
 from prefect.logging import get_logger
 from pydantic import BaseModel
-from vespa.io import VespaQueryResponse, VespaResponse
+from vespa.io import VespaQueryResponse
 
-from flows.deindex import remove_concepts_from_existing_vespa_concepts
 from flows.index import CONCEPT_COUNT_SEPARATOR
 from scripts.cloud import ClassifierSpec
 from src.concept import Concept
-from src.exceptions import PartialUpdateError, QueryError
+from src.exceptions import QueryError
 from src.identifiers import WikibaseID
 from src.labelled_passage import LabelledPassage
 from src.span import Span
@@ -684,35 +683,3 @@ class ConceptModel(BaseModel):
         raise ValueError(
             f"Could not extract concept name from model name '{self.model_name}'"
         )
-
-
-async def partial_update_text_block(
-    text_block_id: TextBlockId,
-    document_import_id: DocumentImportId,
-    concepts: list[VespaConcept],  # A possibly empty list
-    vespa_search_adapter: VespaSearchAdapter,
-) -> None:
-    """Partial update a singular text block and its concepts, if any."""
-    document_passage_id, document_passage = get_document_passage_from_vespa(
-        text_block_id, document_import_id, vespa_search_adapter
-    )
-
-    data_id = get_data_id_from_vespa_hit_id(document_passage_id)
-
-    # FIXME: I think this is incorrect as they are different...
-    serialised_concepts = remove_concepts_from_existing_vespa_concepts(
-        passage=document_passage,
-        concepts_to_remove=concepts,
-    )
-
-    response: VespaResponse = vespa_search_adapter.client.update_data(  # pyright: ignore[reportOptionalMemberAccess]
-        schema="document_passage",
-        namespace="doc_search",
-        data_id=data_id,
-        fields={"concepts": serialised_concepts},
-    )
-
-    if (status_code := response.get_status_code()) != HTTP_OK:
-        raise PartialUpdateError(data_id, status_code)
-
-    return None
